feat: display UPS url in config
