+++ server/helpers.js
@@ -1,0 +1,10 @@
+function addProtocolIfMissing(url) {
+  if (url && !url.startsWith('http')) {
+    return `https://${url}`;
+  }
+  return url;
+}
+
+module.exports = {
+  addProtocolIfMissing
+};
