+++ server/server.js
@@ -5,16 +5,10 @@ const promMid = require('express-prometheus-middleware');
 const { Client, KubeConfig } = require('kubernetes-client');
 const Request = require('kubernetes-client/backends/request');
 const packageJson = require('../package.json');
-const fs = require('fs');
 const { URL } = require('url');
-const {
-  IdentityManagementService,
-  DataSyncService,
-  MobileServicesMap,
-  MobileSecurityService,
-  PushService
-} = require('./mobile-services-info');
 const { updateAppsAndWatch } = require('./appServices');
+const { getServices } = require('./mobile-services');
+const { addProtocolIfMissing } = require('./helpers');
 const mobileClientCRD = require('./mobile-client-crd.json');
 const pushApplicationCRD = require('./push-application-crd.json');
 const androidVariantCRD = require('./android-variant-crd.json');
@@ -35,45 +29,6 @@ app.use(
 
 const port = process.env.PORT || 4000;
 const configPath = process.env.MOBILE_SERVICES_CONFIG_FILE || '/etc/mdc/servicesConfig.json';
-
-const dataSyncService = {
-  type: DataSyncService.type,
-  mobile: true
-};
-
-const addProtocolIfMissing = function(url) {
-  if (url && !url.startsWith('http')) {
-    return `https://${url}`;
-  }
-  return url;
-};
-const DEFAULT_SERVICES = {
-  version: 'dev',
-  components: [
-    {
-      type: IdentityManagementService.type,
-      version: 'latest',
-      host: addProtocolIfMissing(`${process.env.IDM_URL || process.env.OPENSHIFT_HOST}`),
-      mobile: true
-    },
-    {
-      type: PushService.type,
-      host: addProtocolIfMissing(`${process.env.UPS_URL || process.env.OPENSHIFT_HOST}`),
-      version: 'latest',
-      mobile: true
-    },
-    // {
-    //   type: MetricsService.type,
-    //   url: `https://${process.env.METRICS_URL || process.env.OPENSHIFT_HOST}`
-    // }
-    {
-      type: MobileSecurityService.type,
-      host: addProtocolIfMissing(`${process.env.MSS_URL || process.env.OPENSHIFT_HOST}`),
-      version: 'latest',
-      mobile: true
-    }
-  ]
-};
 
 const DEFAULT_NAMESPACE = 'mobile-console';
 
@@ -145,33 +100,6 @@ function getConfigData(req) {
       email: '${userEmail}'
     }
   }; window.SERVER_DATA= { ENABLE_BUILD_TAB: ${enableBuildTab}, DOCS_PREFIX: '${docsPrefix}' };`;
-}
-
-function getServices(servicesConfigPath) {
-  return new Promise(resolve => {
-    if (fs.existsSync(servicesConfigPath)) {
-      return fs.readFile(servicesConfigPath, (err, data) => {
-        if (err) {
-          console.error(`Failed to read service config file ${servicesConfigPath}, mock data will be used`);
-          return resolve(DEFAULT_SERVICES);
-        }
-        return resolve(JSON.parse(data));
-      });
-    }
-    console.warn(`can not find service config file at ${servicesConfigPath}, mock data will be used`);
-    return resolve(DEFAULT_SERVICES);
-  })
-    .then(servicesUrls => servicesUrls.components.concat([dataSyncService]))
-    .then(services => services.filter(s => s.mobile))
-    .then(services =>
-      services.map(service => {
-        const serviceInfo = MobileServicesMap[service.type];
-        if (serviceInfo) {
-          return { ...serviceInfo, ...service };
-        }
-        return service;
-      })
-    );
 }
 
 async function initKubeClient() {
